Add files via upload
Datapack contains basic templating for Bloodlines, Jobs, Guilds, and Cults. Bloodlines replace the base Origins from the mod, Jobs replace the base classes. Guilds are a new addition that will be made accessible in-game on the official Cliffhouse Games servers; default Guild is "Citizen". Cults are another addition to be made accessible in-game; default Cult is "Uninitiated".